Issue #11024 - Added the field for query.
