+++ config/sync/core.entity_form_display.block_content.faceted_taxonomy_widget.default.yml
@@ -4,6 +4,10 @@ status: true
 dependencies:
   config:
     - block_content.type.faceted_taxonomy_widget
+    - field.field.block_content.faceted_taxonomy_widget.field_select_vocabularies
+    - field.field.block_content.faceted_taxonomy_widget.field_show_number_of_posts
+    - field.field.block_content.faceted_taxonomy_widget.field_vocabulary_order_
+    - field.field.block_content.faceted_taxonomy_widget.field_widget_content_type
     - field.field.block_content.faceted_taxonomy_widget.field_widget_title
   module:
     - field_layout
@@ -17,8 +21,33 @@ targetEntityType: block_content
 bundle: faceted_taxonomy_widget
 mode: default
 content:
+  field_select_vocabularies:
+    weight: 28
+    settings: {  }
+    third_party_settings: {  }
+    type: options_select
+    region: content
+  field_show_number_of_posts:
+    weight: 30
+    settings:
+      display_label: true
+    third_party_settings: {  }
+    type: boolean_checkbox
+    region: content
+  field_vocabulary_order_:
+    weight: 29
+    settings: {  }
+    third_party_settings: {  }
+    type: options_select
+    region: content
+  field_widget_content_type:
+    weight: 27
+    settings: {  }
+    third_party_settings: {  }
+    type: options_select
+    region: content
   field_widget_title:
-    weight: 26
+    weight: 1
     settings:
       size: 60
       placeholder: ''
@@ -27,7 +56,7 @@ content:
     region: content
   info:
     type: string_textfield
-    weight: -5
+    weight: 2
     region: content
     settings:
       size: 60
