+++ config/sync/core.extension.yml
@@ -84,6 +84,7 @@ module:
   image_widget_crop: 0
   inline_form_errors: 0
   key_value_field: 0
+  kint: 0
   language: 0
   layout_builder: 0
   layout_discovery: 0
