+++ config/sync/field.field.block_content.faceted_taxonomy_widget.field_show_number_of_posts.yml
@@ -1,0 +1,23 @@
+uuid: 07a3a4ee-8d0b-4728-bed2-8ad9c4b5261a
+langcode: en
+status: true
+dependencies:
+  config:
+    - block_content.type.faceted_taxonomy_widget
+    - field.storage.block_content.field_show_number_of_posts
+id: block_content.faceted_taxonomy_widget.field_show_number_of_posts
+field_name: field_show_number_of_posts
+entity_type: block_content
+bundle: faceted_taxonomy_widget
+label: 'Show number of posts'
+description: 'Show number of posts tagged to a term next to the term.'
+required: false
+translatable: false
+default_value:
+  -
+    value: 0
+default_value_callback: ''
+settings:
+  on_label: 'On'
+  off_label: 'Off'
+field_type: boolean
