+++ config/sync/field.storage.block_content.field_select_vocabularies.yml
@@ -1,0 +1,21 @@
+uuid: 270129a9-b613-478f-8046-eb38c9acc8df
+langcode: en
+status: true
+dependencies:
+  module:
+    - block_content
+    - options
+id: block_content.field_select_vocabularies
+field_name: field_select_vocabularies
+entity_type: block_content
+type: list_string
+settings:
+  allowed_values: {  }
+  allowed_values_function: ''
+module: options
+locked: false
+cardinality: -1
+translatable: true
+indexes: {  }
+persist_with_no_fields: false
+custom_storage: false
