+++ config/sync/field.storage.block_content.field_vocabulary_order_.yml
@@ -1,0 +1,27 @@
+uuid: 22617acd-911c-46a7-a48c-55cf2a11b192
+langcode: en
+status: true
+dependencies:
+  module:
+    - block_content
+    - options
+id: block_content.field_vocabulary_order_
+field_name: field_vocabulary_order_
+entity_type: block_content
+type: list_string
+settings:
+  allowed_values:
+    -
+      value: alphabetical
+      label: Alphabetical
+    -
+      value: weight
+      label: 'Vocab Weight'
+  allowed_values_function: ''
+module: options
+locked: false
+cardinality: 1
+translatable: true
+indexes: {  }
+persist_with_no_fields: false
+custom_storage: false
