+++ config/sync/field.storage.block_content.field_widget_content_type.yml
@@ -1,0 +1,30 @@
+uuid: 22ded049-4895-4f59-bbe0-cea20d4831f3
+langcode: en
+status: true
+dependencies:
+  module:
+    - block_content
+    - options
+id: block_content.field_widget_content_type
+field_name: field_widget_content_type
+entity_type: block_content
+type: list_string
+settings:
+  allowed_values:
+    -
+      value: blog
+      label: Blog
+    -
+      value: page
+      label: Page
+    -
+      value: class
+      label: Class
+  allowed_values_function: ''
+module: options
+locked: false
+cardinality: 1
+translatable: true
+indexes: {  }
+persist_with_no_fields: false
+custom_storage: false
